refactor: move Android record-video orchestration into AndroidBackend
With the video plumbing extracted to SimUseVideo, the Android recording
engine no longer needs to live inline in the SimUse executable target.
Move it into AndroidRecordVideoCommand and register it, giving
record-video the same three-surface layout as every other cross-platform
verb (top-level forwarder + `ios` + `android`); the top-level Android
branch now forwards to AndroidRecordVideoCommand.record(), symmetric to
AndroidScreenshotCommand.performScreenshot.

Flag validation is shared across all three surfaces via
VideoRecordingOptions in SimUseVideo so the contract cannot drift.
Flags and behavior are unchanged.

Verified live against emulator-5554: both `sim-use android record-video`
and top-level `sim-use record-video` produce a valid H.264 MP4 (ffprobe)
with clean SIGTERM finalization.
