+++ Tests/AndroidBackendTests/AndroidRecordVideoArgumentTests.swift
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 import Testing
-@testable import SimUse
+@testable import AndroidBackend
 
 @Suite("Android record-video argument construction")
 struct AndroidRecordVideoArgumentTests {
     @Test("parseWMSize reads Physical size")
     func physicalSize() {
-        let parsed = RecordVideo.parseWMSize("Physical size: 1080x2400\n")
+        let parsed = AndroidRecordVideoCommand.parseWMSize("Physical size: 1080x2400\n")
         #expect(parsed?.width == 1080)
         #expect(parsed?.height == 2400)
     }
@@ -14,25 +14,25 @@ struct AndroidRecordVideoArgumentTests {
     @Test("parseWMSize prefers an Override size over Physical size")
     func overrideSizePreferred() {
         let output = "Physical size: 1080x2400\nOverride size: 720x1600\n"
-        let parsed = RecordVideo.parseWMSize(output)
+        let parsed = AndroidRecordVideoCommand.parseWMSize(output)
         #expect(parsed?.width == 720)
         #expect(parsed?.height == 1600)
     }
 
     @Test("parseWMSize returns nil for unparseable output")
     func unparseable() {
-        #expect(RecordVideo.parseWMSize("cannot connect to display\n") == nil)
+        #expect(AndroidRecordVideoCommand.parseWMSize("cannot connect to display\n") == nil)
     }
 
     @Test("screenrecordArguments omits --time-limit below API 34")
     func api33NoTimeLimit() {
-        let args = RecordVideo.screenrecordArguments(serial: "emu-1", sdk: 33, bitrate: nil, size: nil)
+        let args = AndroidRecordVideoCommand.screenrecordArguments(serial: "emu-1", sdk: 33, bitrate: nil, size: nil)
         #expect(args == ["-s", "emu-1", "exec-out", "screenrecord", "--output-format=h264", "-"])
     }
 
     @Test("screenrecordArguments adds --time-limit 0 on API 34+")
     func api34Unlimited() {
-        let args = RecordVideo.screenrecordArguments(serial: "emu-1", sdk: 34, bitrate: nil, size: nil)
+        let args = AndroidRecordVideoCommand.screenrecordArguments(serial: "emu-1", sdk: 34, bitrate: nil, size: nil)
         #expect(args.contains("--time-limit"))
         #expect(args.contains("0"))
         #expect(args.last == "-")
@@ -40,7 +40,7 @@ struct AndroidRecordVideoArgumentTests {
 
     @Test("screenrecordArguments includes --bit-rate and --size when provided")
     func bitrateAndSize() {
-        let args = RecordVideo.screenrecordArguments(serial: "emu-1", sdk: 34, bitrate: 4_000_000, size: (width: 540, height: 1200))
+        let args = AndroidRecordVideoCommand.screenrecordArguments(serial: "emu-1", sdk: 34, bitrate: 4_000_000, size: (width: 540, height: 1200))
         #expect(args == [
             "-s", "emu-1", "exec-out", "screenrecord", "--output-format=h264",
             "--time-limit", "0",
@@ -57,7 +57,7 @@ struct AndroidRecordVideoArgumentTests {
     // itself is scale-gated.
     @Test("screenrecordArguments passes --bit-rate without --size (default scale)")
     func bitrateWithoutSizeAtDefaultScale() {
-        let args = RecordVideo.screenrecordArguments(serial: "emu-1", sdk: 34, bitrate: 4_000_000, size: nil)
+        let args = AndroidRecordVideoCommand.screenrecordArguments(serial: "emu-1", sdk: 34, bitrate: 4_000_000, size: nil)
         #expect(args.contains("--bit-rate"))
         #expect(args.contains("4000000"))
         #expect(!args.contains("--size"))
@@ -65,14 +65,14 @@ struct AndroidRecordVideoArgumentTests {
 
     @Test("scaledSize halves dimensions and rounds down to even")
     func scaledSizeRounding() {
-        let scaled = RecordVideo.scaledSize((width: 1081, height: 2401), scale: 0.5)
+        let scaled = AndroidRecordVideoCommand.scaledSize((width: 1081, height: 2401), scale: 0.5)
         #expect(scaled.width == 540)
         #expect(scaled.height == 1200)
     }
 
     @Test("scaledSize at 1.0 returns the input unchanged (already even)")
     func scaledSizeIdentity() {
-        let scaled = RecordVideo.scaledSize((width: 1080, height: 2400), scale: 1.0)
+        let scaled = AndroidRecordVideoCommand.scaledSize((width: 1080, height: 2400), scale: 1.0)
         #expect(scaled.width == 1080)
         #expect(scaled.height == 2400)
     }
